tests-requirements.txt
coverage
pytest-benchmark
pylama
tox
